Avoid unnecessary boxing
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@25869 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/smg/SMGTransferRelation.java
@@ -1153,7 +1153,7 @@ public class SMGTransferRelation
             CType innerType = ((CBitFieldType) memberDcl.getType()).getType();
 
             if (memberSize == 0) {
-              offset = machineModel.calculatePaddedBitsize((long) 0, offset, innerType, sizeOfByte);
+              offset = machineModel.calculatePaddedBitsize(0, offset, innerType, sizeOfByte);
             } else {
               offset =
                   machineModel.calculateNecessaryBitfieldOffset(
